+++ src/features/Projects/stores/slices/assets.ts
@@ -37,7 +37,7 @@ export const createAssetsSlice: StateCreator<
 ) => ({
     assetsLastId: 0,
     addAsset(asset) {
-        console.debug("Adding asset", asset);
+        debug(0, "[assets] Adding asset", asset);
         const assets = get().project.assets;
 
         const foundAsset = assets.has(asset.path)
@@ -52,14 +52,7 @@ export const createAssetsSlice: StateCreator<
 
             set({});
         } else {
-            console.debug(
-                "Asset added",
-                asset,
-                loadByAsset(
-                    asset.name,
-                    asset.kind,
-                ),
-            );
+            debug(0, "[assets] Asset added", asset);
 
             assets.set(asset.path, {
                 ...asset,