+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.3.5] - 2025-06-22
+
+### Fixed
+
+- Fix cases where assets weren't loading using `assets/`
+- Fix importing project
+
 ## [2.3.4] - 2025-06-21
 
 ### Fixed